WordCamp Coming Soon Page: Remove Jetpack's Custom CSS stylesheet.
The original intent was to remove it, so that the template would have a consistent base of styles, but it wasn't removed because I assumed that the Custom CSS module was using `wp_enqueue_style()`.

When it's enabled, styles intended for the main site can bleed over to the Coming Soon page, and they don't usually blend well with the Coming Soon design.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@1043 74240141-8908-4e6f-9713-ba540dce6ec7