WordCamp Post Types: Give preference to current page when deciding anchors.
This supports a new use case where the admin sets up a multiple pages that each contain `[schedule]`, `[speakers]`, and `[sessions]`. A camp might want to have page for each day that contained all of the information about what's going on that day, but not what's scheduled for other days.

In that situation, the anchor links in the schedule should point to the current page, rather than searching for the first page that contains the shortcodes.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@1372 74240141-8908-4e6f-9713-ba540dce6ec7