Breathe: Handbook sidebar nav redesign.
The floating sidebar nav was causing readability issues and so I redesigned it to be anchored better and be more evident that it's a sidebar nav. Also redesigned the Topic in-page box to be lighter so it doesn't detract from the content as much. Still working on a better solution.

Props grapplerulrich

See #2389, #2410


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@4758 74240141-8908-4e6f-9713-ba540dce6ec7