Official WordPress Events: Only include groups where the WP user is organizer
There are a handful of groups that the WordPress user has joined, but has no role in, or has the `Co-Organizer` role. Those groups haven't joined the chapter program, or have started the process of joining, but not completed it. The intent of this plugin -- and everything that pulls data from it -- is to only promote events that are part of the chapter program.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@5592 74240141-8908-4e6f-9713-ba540dce6ec7