WordCamp Sponsors: Additional parameters for the shortcode
This adds or updates the following parameters in the `[sponsors]`
shortcode:

**link**

The default value `none` will not link the title and logo. The current
value `website` will link both to the external sponsor website. The new
value `post` will link both to the single sponsors post.

**title**

The default value `visible` will show the h3 with the sponsors title,
linked accordingly to the link attribute. Other values will omit the
title.

**content**

The default value `full` will print the content. The value `excerpt`
will print the first 55 words of the content, adding a "continue
reading" (with hidden screen reader title) at the end. Any other value,
like `none`, will hide the content, so only the logo and/or title will
be shown.

**excerpt_length**

Sets the number of words to be used in the excerpt (default is 55).

Props Kau-Boy
Fixes #1751

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@5854 74240141-8908-4e6f-9713-ba540dce6ec7