Meetup Client: Throttle immediately after making a request.
This ensures that it always gets called after making a request, which avoids edge cases where the script needs to pause before making a new request, but wouldn't, because it was done with the current batch of requests, and unaware that another batch would be started immediately after.

This also makes it easier to maintain, since it's only in one place, rather than us having to remember to add it to all of the calling functions.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@6915 74240141-8908-4e6f-9713-ba540dce6ec7