WP15: Unset referenced `$event` after loop to avoid duplicates.
After looping through an array by reference, the final value remains a reference unless it's unset. If you then reuse that variable in a future loop, it results in one of the items being overwritten by another. In this case, that meant that the `Arnhem` event was displayed twice, but the `New Delhi` was skipped.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@6996 74240141-8908-4e6f-9713-ba540dce6ec7