Events: Only restrict results to a country when explicitly requested.
Otherwise, this creates a situation where an event within driving distance is not shown to somebody, just because it's on the other side of a border. There are some cases where the event might not be relevant (e.g., different languages are spoken in those countries), but there are many others where the event is relevant, so it's better to err on the side of showing too many events than too few.

See https://wordpress.slack.com/archives/C08M59V3P/p1524168308000202.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@7145 74240141-8908-4e6f-9713-ba540dce6ec7