CampTix-Invoices: Fix path for custom logo to include it in PDF
Currently if a custom logo is set for the invoice, it's not included in the generated PDF.

In testing locally, it can be observed  that the default logo works and is pulled in as a /absolute/path/to/file.png where as custom logos are https://wordcamp.site/path/to/file.png.

The attached patch attempts to use absolute path to the image in the hope that's the problem, invoice generation works before and after for the test environment.

props dion

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8798 74240141-8908-4e6f-9713-ba540dce6ec7