Slack: Announcements: Switch `/announce` and `/here` to use an at-here reserving at-channel for a new `/at-channel` command.
Due to the number of users in this slack instance now, an at-channel is not usually the intended default and causes notifications for a lot of users it's not necessarily relevant to.

`/at-channel` is still limited to those who have been specifically granted access, which is the same as those who could previously use `/announce` and `/here`..

This also includes some line breaks and rewording of the descriptive texts.

See https://wordpress.slack.com/archives/C02RQBWTW/p1604090126182300.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10423 74240141-8908-4e6f-9713-ba540dce6ec7