Translate: Add a warning that triggers on unexpected malformed sptintf placeholders.
This is to handle the cases where a literal `percent` word is replaced by `%` within a printf-string without being escaped, as has happened in learn.wordpress.org strings multiple times.

Included is a translation fixer, which automatically escapes the % to %% when possible, as long as the resulting string passes through the other "placeholders match" checks.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10946 74240141-8908-4e6f-9713-ba540dce6ec7