Translate: Simplify translations table.
The priority column is most of the time empty and as such a waste of space which is better used to enhance the display of originals and translations. The priority column is now hidden and if an original is marked as high or low priority it is displayed right next to the original similiar to the existing context bubble.
The context is now also hidden if it doesn't provide any new information to translators. This is the mostly case when the context is the same or a cut-off version of the original.

Fixes #5778.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11051 74240141-8908-4e6f-9713-ba540dce6ec7