Plugin Directory: Remove "style" requirement when validating block plugins.
Previously, either `editorStyle` or `style` was required for block plugins, but this was not communicated to the plugin developer (only as a recommendation). This caused some block plugins to pass the validation in UI, but still fail validation behind the scenes. Omitting a style doesn't break anything on the core side, so it should be allowed.

Props oskosk for flagging an affected block.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11096 74240141-8908-4e6f-9713-ba540dce6ec7