Plugin Directory: When release confirmations are enabled, always build the trunk `plugin-name.zip` zip.
When release confirmations are enabled, trunk can never be set as the stable tag, so it's always safe to build the trunk zip.

There's still a limitation here - If a commit introduces a new stable release AND modifies trunk in the same commit, the trunk ZIP won't be updated until the release is confirmed.
It'd be nice to still rebuild the ZIP in that case, so a follow up change may still be required.

See #5352.
See #5901.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11238 74240141-8908-4e6f-9713-ba540dce6ec7