Profiles: Display GitHub activity on profiles.
This is implemented through a WordPress GitHub organization webhook, which parses pushed events and stores it in a table for later querying.
This is how Trac/SVN (and now GitHub) activity is inserted into the profiles activity stream.

Issues created & closed, Pull Requests submitted & merged, and pushes to repositories are supported.

Activity is only recorded against a WordPress.org profile if their GitHub profile is linked to WordPress.org, but will show up once that's done.

This is still a work in progress, and we may want to import older history as well.

See #4447, #5592.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11270 74240141-8908-4e6f-9713-ba540dce6ec7