Profiles: Remove unnecessary Learn handler abstraction.
The client is already authenticated, and is already passing `activity_args`, so the abstraction just adds a lot of boilerplate that Learn (and most other handlers) don't need.

`handle_activity()` can still act as a middleman to preserve backcompat if BuddyPress ever changes the options, but that's unlikely. `format_callback` can also be used if we ever want to override the `action` passed by the client.

See https://github.com/WordPress/five-for-the-future/issues/191

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11829 74240141-8908-4e6f-9713-ba540dce6ec7