Patterns API: Update pagination & sort order for fetching patterns.
Use the `per_page` parameter passed in if it exists, otherwise continue defaulting to 100 patterns. Remove the forced alphabetical title order in favor of the value passed in. Patterns will now be returned in the same order they are viewed on the website. `orderby=favorite_count` can be used to mirror the "popular" sorting.

Fixes https://github.com/WordPress/pattern-directory/issues/529.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@12147 74240141-8908-4e6f-9713-ba540dce6ec7