Plugin Directory: Import: Update the 'tags' post meta last. This ensures that there's no public links to a ZIP before it's created.
When a plugin is released, the slowest part of the processing is checking it out from SVN and building the ZIP files.
If a ZIP url is accessed before the ZIP has been built, the non-existence of that ZIP may be cached, causing problems for downloads.

See https://wordpress.slack.com/archives/C02QB8GMM/p1677722996478089?thread_ts=1677683016.024589&cid=C02QB8GMM


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@12438 74240141-8908-4e6f-9713-ba540dce6ec7