Support Forums: Show context information about pending posts.
When viewing queues of pending posts, a moderator is expected to check a users history to make sure they un-flag users in as timely a manner as possible. Doing so generally requires going in and checking the users posting history manually via their profiles.

This change reduces the workload of moderators by providing a lot of this information within context of a pending post. 

Fixes #6809.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@12616 74240141-8908-4e6f-9713-ba540dce6ec7